esp: a timid attempt at an async command queue
The IR sending command was already async by design to not block during
the whole 100ms long bursts. Now when two or more commands are sent
subsequently they're put in a queue and executed as soon as the current
one is completed.

The whole thing is poorly designed, barely tested and can probably fail
in several nasty ways. I like it nonetheless.