esp: add power off mode
Power off is actually just a nibble in the whole command. All the state
still works while in off mode (only the display status is actually
visible to the user). I think they use this state for scheduled turn on.

For our purpose the only desired outcome is turning off the AC so the
whole state doesn't really matter (except display) as we're sending it
again on power on anyway.